chore(prometheus): infer procs and mem from resources spec (#198)
* chore(prometheus): infer procs and mem from resources spec

This will set GOMAXPROCS and GOMEMLIMIT to match the limits set configured spec, which is the recommended Kubernetes-native way.

CFS throttling is enforced by measuring your usage against your limits, and the big idea behind reducing throttling is to set your GOMAXPROCS according to your CPU limit - this leads to OOMKilled for pods

In Kubernetes, GOMAXPROCS defaults to the number of CPU cores on the Node, not the Pod. If your Pod has a much lower CPU limit (e.g., 1 core on a 32-core node), your Go app will still try to run with GOMAXPROCS=32 . This mismatch can hurt performance due to unnecessary context switching and CPU contention.

* chore: disable dupliacte mimir push

- Alloy will push metrics to both prometheus and mimir